move calc to js directory, change import statement in test file
